Fix _make_kde bug but skl method need to be check later
